pyproject.toml
[tool.poetry]
name = "into-v2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "into"}]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.2"
fastapi = "^0.90.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
psycopg2-binary = "^2.9.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.9.3"
tenacity = "^8.2.0"
requests = "^2.28.2"
click = "^8.1.3"
aioredis = "^2.0.1"
pyjwt = "^2.6.0"
aiohttp = "^3.8.4"
alibabacloud-cloudauth20190307 = "2.0.3"
asyncpg = "^0.27.0"
celery = "^5.2.7"
redis = "^4.5.1"
flower = "^1.2.0"
web3 = "^6.0.0"
black = "^23.1.0"
cos-python-sdk-v5 = "^1.9.23"
python-multipart = "^0.0.6"
tweepy = "^4.13.0"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
